msac: Add a cast to indicate intended narrowing from size_t to unsigned
This fixes this compiler warning with MSVC:
../src/msac.c(148): warning C4267: '+=': conversion from 'size_t' to 'unsigned int', possible loss of data

+++ src/msac.c
@@ -145,7 +145,7 @@ static unsigned decode_symbol(MsacContext *const s, const uint16_t *const cdf,
         u = v;
         v = r * (cdf[ret++] >> EC_PROB_SHIFT);
         v >>= 7 - EC_PROB_SHIFT;
-        v += EC_MIN_PROB * (n_symbols - ret);
+        v += EC_MIN_PROB * (int) (n_symbols - ret);
     } while (c < v);
 
     assert(u <= s->rng);
